analitics/requirements.txt
boto3==1.20.46
statistics==1.0.3.5
python-dotenv==0.19.2
pytest==2.9.1
requests==2.27.1
urllib3
sockets
paramiko
setuptools_rust
subprocess
parse==1.19.0